requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.0
Django==2.2.1
django-cors-headers==3.0.0
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.7.1
pytz==2018.4
requests==2.21.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
tencentcloud-sdk-python==3.0.68
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==16.0.0
wcwidth==0.1.7
